Update nvim config

+++ .config/nvim/template/cpp/base-atcoder.cpp
@@ -1,6 +1,6 @@
 #include <bits/stdc++.h>
 using namespace std;
-#define rep(i, n) for (int i = 0; i < (int)(n); i++)
+#define rep(i, n) for (ll i = 0; i < (ll)(n); i++)
 using ll = long long;
 
 int main() {
